+++ packages/cli/src/commands/sandbox/sandbox-delete/sandbox_delete_command.ts
@@ -45,7 +45,7 @@ export class SandboxDeleteCommand
     if (isConfirmed) {
       await (
         await this.sandboxFactory.getInstance()
-      ).delete({ identifier: args.identifier });
+      ).delete({ identifier: args.identifier, profile: args.profile });
     }
   };