+++ packages/client-config/src/client-config-writer/client_config_to_legacy_converter.test.ts
@@ -25,7 +25,7 @@ void describe('ClientConfigLegacyConverter', () => {
           // eslint-disable-next-line @typescript-eslint/no-explicit-any
           version: '3' as any,
         }),
-      new AmplifyFault('UnsupportedClientConfigVersion', {
+      new AmplifyFault('UnsupportedClientConfigVersionFault', {
         message: 'Only version 1 of ClientConfig is supported.',
       })
     );