+++ packages/backend-data/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @aws-amplify/backend-data
 
+## 1.1.2
+
+### Patch Changes
+
+- 4d4c0d5: Use proper error and fault suffixes.
+
+## 1.1.1
+
+### Patch Changes
+
+- a65371c: upgrade aws-cdk and aws-cdk-lib to ^2.152.0
+- Updated dependencies [a65371c]
+  - @aws-amplify/backend-output-storage@1.1.1
+  - @aws-amplify/plugin-types@1.2.1
+
 ## 1.1.0
 
 ### Minor Changes