+++ packages/backend-function/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @aws-amplify/backend-function
 
+## 1.3.4
+
+### Patch Changes
+
+- 67feca0: update env type cast to cast as unknown first
+
+## 1.3.3
+
+### Patch Changes
+
+- 775e2bb: Fix InvalidFunctionConfigurationError and InvalidStorageAccessDefinitionError error names.
+
+## 1.3.2
+
+### Patch Changes
+
+- a65371c: upgrade aws-cdk and aws-cdk-lib to ^2.152.0
+- Updated dependencies [a65371c]
+  - @aws-amplify/backend-output-storage@1.1.1
+  - @aws-amplify/plugin-types@1.2.1
+
 ## 1.3.1
 
 ### Patch Changes