+++ packages/client-config/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @aws-amplify/client-config
 
+## 1.2.1
+
+### Patch Changes
+
+- 4d4c0d5: Use proper error and fault suffixes.
+- Updated dependencies [4d4c0d5]
+  - @aws-amplify/model-generator@1.0.5
+
+## 1.2.0
+
+### Minor Changes
+
+- d9b83a1: support adding more than one bucket
+
+### Patch Changes
+
+- Updated dependencies [d9b83a1]
+  - @aws-amplify/deployed-backend-client@1.4.0
+  - @aws-amplify/backend-output-schemas@1.2.0
+
 ## 1.1.3
 
 ### Patch Changes